Fix signedness issues

+++ include/Exceptions/Bad_Still_Alive_Line.h
@@ -8,7 +8,7 @@
 class Bad_Still_Alive_Line : public std::runtime_error
 {
     public:
-        Bad_Still_Alive_Line(int line_number, std::string& line);
+        Bad_Still_Alive_Line(size_t line_number, std::string& line);
 };
 
 #endif // BAD_STILL_ALIVE_LINE_H
